+++ frontend/app/src/components/home/ChatEventList.svelte
@@ -538,7 +538,7 @@
                     pathState.route.kind === "selected_channel_route") &&
                 (pathState.route.open || pathState.route.threadMessageIndex !== undefined)
             ) {
-                client.setFocusThreadMessageIndex(chat.id, pathState.route.threadMessageIndex);
+                client.setFocusThreadMessageIndex(pathState.route.threadMessageIndex);
                 client.openThread(msgEvent, false);
             } else {
                 client.closeThread();
@@ -723,9 +723,9 @@
 
     function setFocusMessageIndex(messageIndex: number | undefined) {
         if (threadRootEvent === undefined) {
-            client.setFocusMessageIndex(chat.id, messageIndex);
+            client.setFocusMessageIndex(messageIndex);
         } else {
-            client.setFocusThreadMessageIndex(chat.id, messageIndex);
+            client.setFocusThreadMessageIndex(messageIndex);
         }
     }
 </script>
